+++ apps/_root/app/layout.tsx
@@ -4,6 +4,21 @@ import '@sushiswap/ui/index.css'
 
 import React from 'react'
 import { Providers } from './providers'
+import { SanctionedAddressDialog } from '@sushiswap/wagmi/components/SanctionedAddressDialog'
+
+import { Inter, Roboto_Mono } from 'next/font/google'
+
+const inter = Inter({
+  subsets: ['latin'],
+  display: 'swap',
+  variable: '--font-inter',
+})
+
+const roboto_mono = Roboto_Mono({
+  subsets: ['latin'],
+  display: 'swap',
+  variable: '--font-roboto-mono',
+})
 
 export const metadata: Metadata = {
   title: {
@@ -17,7 +32,7 @@ export const metadata: Metadata = {
 export default function RootLayout({ children }: { children: React.ReactNode }) {
   return (
     // <html lang="en" className="[color-scheme:dark]">
-    <html lang="en" className="dark">
+    <html lang="en" className={`${inter.variable} ${roboto_mono.variable} dark`}>
       <link rel="apple-touch-icon" sizes="180x180" href="/apple-touch-icon.png?v=1" />
       <link rel="icon" type="image/png" sizes="32x32" href="/favicon-32x32.png?v=1" />
       <link rel="icon" type="image/png" sizes="16x16" href="/favicon-16x16.png?v=1" />
@@ -25,7 +40,10 @@ export default function RootLayout({ children }: { children: React.ReactNode }) 
       <link rel="mask-icon" href="/safari-pinned-tab.svg?v=1" color="#fa52a0" />
       <link rel="shortcut icon" href="/favicon.ico?v=1" />
       <body className="h-screen">
-        <Providers>{children}</Providers>
+        <Providers>
+          <SanctionedAddressDialog />
+          {children}
+        </Providers>
       </body>
     </html>
   )