+++ apps/academy/pages/products/furo.tsx
@@ -26,7 +26,7 @@ const { color, cards, buttonText, productStats, faq } = PRODUCTS_DATA[PRODUCT_SL
 export const getStaticProps: GetStaticProps = async () => {
   const data = await getProducts({ filters: { slug: { eq: PRODUCT_SLUG } } })
   const product = data?.products?.data?.[0].attributes
-  if (!product) throw new Error(`Product not found`)
+  if (!product) throw new Error('Product not found')
   return { props: product, revalidate: 60 }
 }
 
@@ -34,7 +34,7 @@ const ProductPage: FC<InferGetStaticPropsType<typeof getStaticProps>> = (product
   const { name, longName, url, description, slug, relevantArticleIds } = product
 
   const { data, isValidating } = useSWR(
-    [`/furo-articles`],
+    ['/furo-articles'],
     async () => await getLatestAndRelevantArticles(slug, relevantArticleIds),
     {
       revalidateOnFocus: false,